Alterado Latitude e Longitude para Single

+++ API.XF.VKNGNotes/Models/Note.cs
@@ -10,8 +10,8 @@ namespace API.XF.VKNGNotes.Models
         public int Id { get; set; }
         public string Titulo { get; set; }
         public string Detalhes { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public Single Latitude { get; set; }
+        public Single Longitude { get; set; }
 
     }
 }
